refactor use of dirty image in inversion
